Fixes for stringUtils issues * tS macro from stringUtils is used in various user projects compiled in C++98 mode meaning that new stringUtils fail to compile - therefore moved more code-gen specific stuff out into 'codeGenUtils' * Fixed more warnings * Removed more tS usage from GeNN library

+++ lib/src/newNeuronModels.cc
@@ -48,7 +48,7 @@ std::vector<std::pair<std::string, NeuronModels::Base::DerivedParamFunc>> Neuron
     derivedParams.reserve(nm.dpNames.size());
 
     // Loop through derived parameters
-    for(int p = 0; p < nm.dpNames.size(); p++)
+    for(size_t p = 0; p < nm.dpNames.size(); p++)
     {
         // Add pair consisting of parameter name and lambda function which calls
         // through to the DPS object associated with the legacy model
@@ -86,7 +86,7 @@ std::vector<std::pair<std::string, std::string>> NeuronModels::LegacyWrapper::Zi
     zip.reserve(a.size());
 
     // Build vector from legacy neuron model
-    for(int v = 0; v < a.size(); v++)
+    for(size_t v = 0; v < a.size(); v++)
     {
         zip.push_back(std::pair<std::string, std::string>(a[v], b[v]));
     }
